refactor: rename modeling_kimi_k3_linear back to modeling_kimi_linear, remove formatting-only comments
- Rename modeling_kimi_k3_linear.py back to modeling_kimi_linear.py to
  match the upstream HuggingFace filename
- Remove the re-export shim that was in modeling_kimi_linear.py
- Remove "Differences: formatting only" comments from files with no
  substantive changes
