+++ src/llmcompressor/modeling/kimi_k3/configuration_kimi_k3.py
@@ -1,6 +1,3 @@
-# Differences from https://huggingface.co/moonshotai/Kimi-K3/tree/main:
-# - Formatting only (no functional changes)
-
 from typing import Optional
 
 from transformers.configuration_utils import PretrainedConfig
