+++ src/llmcompressor/modeling/kimi_k3/kimi_k3_vision_processing.py
@@ -1,6 +1,3 @@
-# Differences from https://huggingface.co/moonshotai/Kimi-K3/tree/main:
-# - Formatting only (no functional changes)
-
 """Image processor class for Kimi-K3."""
 
 import json
