+++ src/llmcompressor/modeling/kimi_k3/media_utils.py
@@ -1,6 +1,3 @@
-# Differences from https://huggingface.co/moonshotai/Kimi-K3/tree/main:
-# - Formatting only (no functional changes)
-
 import base64
 import functools
 import io
